replace GetWavSize with GetAudioSize using miniaudio to calculate the number of pcm frames
